receiver/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-crontab==0.7.1
idna==2.10
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
thecampy==2.0.0
urllib3==1.26.4
